refactor: apply new pipline to auto chart exclude config pannel (#81)
